adicionando trativas de erro e uso do userId

+++ backend/src/models/pokemonModel.js
@@ -8,86 +8,84 @@ class PokemonModel {
       connection.query(sql, {}, (err, res) => {
         if (err) {
           console.error('Deu erro no listar...', err);
-          reject(err);
+          return reject(new Error(err));
         }
-  
-        console.log('Show...');
-        resolve(res);
+
+        return resolve(res);
       })
     });
   }
 
-  getFavorited(pokeId) {
-    const sql = `SELECT * FROM pokemons WHERE pokeId = ?`;
+  getFavorited(pokeId, userId) {
+    const sql = `SELECT * FROM pokemons WHERE pokeId = ? AND userId = ?`;
   
     return new Promise((resolve, reject) => {
-      connection.query(sql, [pokeId], (err, res) => {
+      connection.query(sql, [pokeId, userId], (err, res) => {
         if (err) {
           console.error('Erro ao buscar Pokémon:', err);
-          reject(err);
+          return reject(new Error(err));
         }
   
-        resolve(res.length);
+        return resolve(res.length > 0);
       });
     });
   }
 
-  post(poke) {
+  post(poke, userId) {
     const { pokeTypes } = poke;
     const sql = `INSERT INTO pokemons SET ?`;
   
     return new Promise((resolve, reject) => {
-      connection.query(sql, { ...poke, pokeTypes: JSON.stringify(pokeTypes) }, (err, res) => {
+      connection.query(sql, { ...poke, pokeTypes: JSON.stringify(pokeTypes), userId }, (err, res) => {
         if (err) {
           console.error('Erro ao inserir Pokémon:', err);
-          reject(err);
+          return reject(new Error(err));
         }
   
         console.log('Pokémon inserido com sucesso.');
         // Retorna o ID inserido e os dados do Pokémon
-        resolve({ id: res.insertId, ...poke });
+        return resolve({ id: res.insertId, ...poke });
       });
     });
   }
 
-  put(poke) {
+  put(poke, userId) {
     const { pokeId, pokeTypes } = poke;
-    const sql = `UPDATE pokemons SET ? WHERE pokeId = ?`;
+    const sql = `UPDATE pokemons SET ? WHERE pokeId = ? AND userId = ?`;
   
     return new Promise((resolve, reject) => {
-      connection.query(sql, [{ ...poke, pokeTypes: JSON.stringify(pokeTypes) }, pokeId], (err, res) => {
+      connection.query(sql, [{ ...poke, pokeTypes: JSON.stringify(pokeTypes) }, pokeId, userId], (err, res) => {
         if (err) {
-          console.error('Erro ao atualizar Pokémon: ', err);
-          reject(err);
+          return reject(new Error(err));
         }
   
         if (res.affectedRows > 0) {
           console.log('Pokémon atualizado com sucesso.');
-          resolve({ updated: true, ...poke });
+          return resolve({ updated: true, ...poke });
         } else {
           console.error('Nenhum registro foi alterado');
-          reject({ updated: false, ...poke });
+          return reject({ updated: false, ...poke });
         }
       });
     });
   }
 
-  delete(pokeId){
-    const sql = `DELETE FROM pokemons WHERE pokeId = ?`;
+  delete(pokeId, userId){
+    const sql = `DELETE FROM pokemons WHERE pokeId = ? AND userId = ?`;
 
     return new Promise((resolve, reject) => {
-      connection.query(sql, [pokeId], (err, res) => {
+      connection.query(sql, [pokeId, userId], (err, res) => {
         if (err) {
           console.error('Erro ao deletar pokemon: ', err)
-          reject(err);
+          return reject(new Error(err));
         }
 
         if (res.affectedRows > 0) {
           console.log('Pokémon excluído com sucesso.');
-          resolve({ deleted: true });
+          return resolve({ deleted: true });
         } else {
           console.log('Pokémon não encontrado.');
-          resolve({ deleted: false });
+          return resolve({ deleted: false });
         }
       });
     });
